refactor:  refactor post layout structure and add X button in goal section (#153)
