+++ src/features/post/components/index.ts
@@ -1,6 +1,7 @@
 export { PostFormLayout } from './PostFormLayout'
 export { PostFormSkeleton } from './PostFormSkeleton'
 export { PostGoalSection } from './PostGoalSection'
+export { PostImageSection } from './PostImageSection'
 export { PostTagSection } from './PostTagSection'
 export { PostVoteSection } from './PostVoteSection'
 export { SectionLabel } from './SectionLabel'
